+++ CHANGELOG.md
@@ -39,6 +39,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Improvements
 
+- [#761](https://github.com/babylonlabs-io/babylon/pull/746) Add mainnet allowed transaction hash to v1 upgrade handler.
+- [#760](https://github.com/babylonlabs-io/babylon/pull/760) Add mainnet BTC headers
+height from `854785` to `890123`.
 - [#757](https://github.com/babylonlabs-io/babylon/pull/757) Statically link wasm and add binaries
 with [`testnet`, `mainnet`] flags to release assets.
 - [#746](https://github.com/babylonlabs-io/babylon/pull/746) Add mainnet parameters to v1 upgrade handler.